requirements.txt
opencv-python>=4.5.5.64
deepface>=0.0.93
tensorflow>=2.19.0
tf-keras>=2.19.0
numpy>=1.14.0
pandas>=0.23.4
Pillow>=5.2.0
Flask>=2.3.0
flask-cors>=4.0.0